files updated due to moving custom widgets to separate folder
